src/raspberryPi/requirements.txt
boto3==1.34.35
botocore==1.34.162
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.8
jmespath==1.0.1
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
RPi.GPIO==0.7.1
s3transfer==0.10.2
six==1.16.0
urllib3==2.2.2